### 2023-10-19 #5960 @editorialbot
[PRE REVIEW]: flowerMD: Flexible Library of Organic Workflows and Extensible Recipes for Molecular Dynamics
**Submitting author:** @marjanalbouye (<a href="http://orcid.org/0009-0001-9565-3076">Marjan Albooyeh</a>)
**Repository:** https://github.com/cmelab/flowerMD
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
