### 2023-10-19 #5960 @Kevin-Mattheus-Moerman
@marjanAlbouye thanks for this submission. I am the EiC for this track and here to help with initial steps. I will start looking for a handling editor. For the moment can you try work on the following:

- [ ] Can you please try to address the above :point_up: potentially missing DOI's
- [ ] In your affiliations, can you please spell out USA as United States of America
- [ ] Please add clear contributing guidelines to your repository. We recommend adding a `CONTRIBUTING.md` file and to link to it in your readme. Here is a useful site showing examples of such a file: https://contributing.md/example/
