+++ paper/paper.bib
@@ -80,19 +80,15 @@
 	langid = {english},
 }
 
-@misc{hoomd_2019,
-	title = {{HOOMD}-blue: A Python package for high-performance molecular dynamics and hard particle Monte Carlo simulations},
-	url = {http://arxiv.org/abs/1308.5587},
-	shorttitle = {{HOOMD}-blue},
-	number = {{arXiv}:1308.5587},
-	publisher = {{arXiv}},
-	author = {Anderson, Joshua A. and Glaser, Jens and Glotzer, Sharon C.},
-	urldate = {2023-10-03},
-	date = {2019-10-18},
-	langid = {english},
-	eprinttype = {arxiv},
-	eprint = {1308.5587 [physics]},
-	keywords = {Physics - Computational Physics},
+@article{anderson2020hoomd,
+  title={HOOMD-blue: A Python package for high-performance molecular dynamics and hard particle Monte Carlo simulations},
+  author={Anderson, Joshua A and Glaser, Jens and Glotzer, Sharon C},
+  journal={Computational Materials Science},
+  volume={173},
+  pages={109363},
+  year={2020},
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2019.109363}
 }
 
 @article{polyply_2022,
@@ -144,7 +140,8 @@
   number={8},
   pages={2906--2922},
   year={2018},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acs.macromol.8b00011}
 }
 
 @Inbook{Klein2016mBuild,
