### 2023-10-23 #5960 @lucydot
Hi @LIVazquezS @csadorf @rosecers - are you available to review this submission to JOSS? It is molecular dynamics + Python + reproducible workflows, which appears in-line with some of your interests...
