### 2023-10-27 #5989 @erjank
@lucydot , a possible perceived COI is that @csadorf and I had the same PhD advisor. I graduated from that lab in 2012, years before @csadorf joined and we have not worked together.
