### 2023-10-30 d18a194
fix write mode in system.to_gsd
