### 2023-10-30 a48260c
add pickle method to system, add unit tests
