+++ flowermd/tests/base/test_system.py
@@ -865,3 +865,35 @@ class TestSystem(BaseTest):
         system.to_gsd("test.gsd")
         assert os.path.isfile(os.path.join(os.getcwd(), "test.gsd")) 
         os.remove(os.path.join(os.getcwd(), "test.gsd")) 
+
+    def test_to_gsd_no_ff(self, polyethylene):
+        polyethylene = polyethylene(lengths=5, num_mols=1)
+        system = Pack(
+            molecules=[polyethylene],
+            density=1.0,
+        )
+        system.to_gsd("test.gsd")
+        assert os.path.isfile(os.path.join(os.getcwd(), "test.gsd")) 
+        os.remove(os.path.join(os.getcwd(), "test.gsd")) 
+
+    def test_pickle_ff(self, polyethylene):
+        polyethylene = polyethylene(lengths=5, num_mols=1)
+        system = Pack(
+            molecules=[polyethylene],
+            density=1.0,
+        )
+        system.apply_forcefield(
+            r_cut=2.5, force_field=[OPLS_AA()], auto_scale=False
+        )
+        system.pickle_forcefield("forcefield.pickle")
+        assert os.path.isfile(os.path.join(os.getcwd(), "forcefield.pickle")) 
+        os.remove(os.path.join(os.getcwd(), "forcefield.pickle")) 
+
+    def test_pickle_ff_not_ff(self, polyethylene):
+        polyethylene = polyethylene(lengths=5, num_mols=1)
+        system = Pack(
+            molecules=[polyethylene],
+            density=1.0,
+        )
+        with pytest.raises(ValueError):
+            system.pickle_forcefield("forcefield.pickle")
