### 2023-10-30 c99e65f
update doc strings
