+++ flowermd/base/system.py
@@ -406,9 +406,9 @@ class System(ABC):
     def pickle_forcefield(self, file_path="forcefield.pickle"):
         """Pickle the list of HOOMD forces.
 
-        This method useful for saving the forcefield of a simulation to a file
-        and reusing it for restarting a simulation or running a different
-        simulation.
+        This method useful for saving the Hoomd force objects
+        generated by System.apply_forcefield() which can then
+        be used to initialize a Simulation at a later time.
 
         Parameters
         ----------
