### 2023-10-30 d4c3ab7
add flush method for writers, add max write buffer size param in Simulation
