+++ flowermd/tests/base/test_simulation.py
@@ -359,8 +359,9 @@ class TestSimulate(BaseTest):
         os.remove("sim_data.txt")
 
     def test_flush(self, benzene_system):
-        sim = Simulation.from_system(benzene_system)
-        sim.run_NVT(kT=1.0, tau_kt=0.01, n_steps=500)
+        sim = Simulation.from_system(benzene_system, gsd_write_freq=100)
+        sim.run_NVT(kT=1.0, tau_kt=0.01, n_steps=500, write_at_start=False)
         sim.flush_writers()
-        assert os.path.isfile("trajectory.gsd")
+        with gsd.hoomd.open("trajectory.gsd") as traj:
+            assert len(traj) > 0
         os.remove("trajectory.gsd")
