### 2023-11-09 f2fa27b
create EllipsoidChain class
