### 2023-11-11 #5989 @abhishektiwari
@mbarzegary Please see my initial comments.

## General checks

### Contribution and authorship
Can you provide more details on Eric Jankowski contributions to paper or software?

## Software paper

### Reference
Incorrect reference to HOOMD-blue paper at line 8 and 28, should be pointing to `anderson2020hoomd` as included in `paper.bib` file.

### Quality of writing
A few typos

— `macromolecular` not `macrmolecular`
— `because` not `becuase`

## Suggested improvements
— A lot of deprecation warnings due to use of older `numpy` interfaces. Worth making sure there is path to latest behaviour. E.g. — Conversion of an array with ndim > 0 to a scalar is deprecated, and will error in future.
—  Provide `pip` package to make installation. I believe authors already working [on it.](https://github.com/cmelab/flowerMD/issues/13)
— There is potential to use directed acyclic graphs (DAG) execution for workflow recipes. Worth exploring a few examples using Airflow or likes.

## Overall
I truly enjoyed reviewing this paper. Everything from installation to functional testing to tutorial, I was able to finish within few hours. Kudos to authors for providing a very frictionless experience. Tutorial cover variety of well documented use cases. Package APIs are well documented and include type hints at parameters level. More importantly, package is structured in a way which encourage extensibility.
