### 2023-11-13 c31a4ac
update env file to use conda packages for cmeutils and grits, pin latest versions
