+++ environment-gpu.yml
@@ -2,13 +2,13 @@ name: flowermd
 channels:
   - conda-forge
 dependencies:
-  - boltons
   - foyer
   - fresnel
-  - freud>=2.13.1
-  - gsd>=3.0
-  - hoomd=4.1=*gpu*
-  - mbuild
+  - freud >=2.13.1
+  - gmso >=0.11.2
+  - gsd >=3.0
+  - hoomd=4.3=*gpu*
+  - mbuild >=0.16.4
   - numpy
   - openbabel
   - pip
@@ -16,12 +16,6 @@ dependencies:
   - pytest
   - pytest-cov
   - python=3.9
-  - unyt
-  - symengine
-  - python-symengine
-  - sympy
-  - fresnel
-  - pip:
-      - git+https://github.com/mosdef-hub/gmso.git@main
-      - git+https://github.com/cmelab/cmeutils.git@master
-      - git+https://github.com/cmelab/grits.git@main
+  - fresnel >=0.13.5
+  - cmeutils >=0.0.3
+  - grits >=0.3.0
