+++ flowermd/library/surfaces.py
@@ -1,10 +1,13 @@
 """Recipes to generate surfaces using mBuild."""
 
+import mbuild as mb
 from mbuild.compound import Compound
 from mbuild.lattice import Lattice
 
+from flowermd.base import Molecule
 
-class Graphene(Compound):
+
+class Graphene(Molecule):
     """Create a rectangular graphene layer or multiple layers.
 
     Parameters
@@ -28,9 +31,14 @@ class Graphene(Compound):
     """
 
     def __init__(
-        self, x_repeat, y_repeat, n_layers, periodicity=(True, True, False)
+        self,
+        x_repeat,
+        y_repeat,
+        n_layers,
+        force_field=None,
+        periodicity=(True, True, False),
     ):
-        super(Graphene, self).__init__(periodicity=periodicity)
+        surface = mb.Compound(periodicity=periodicity)
         spacings = [0.425, 0.246, 0.35]
         points = [[1 / 6, 0, 0], [1 / 2, 0, 0], [0, 0.5, 0], [2 / 3, 1 / 2, 0]]
         lattice = Lattice(
@@ -38,9 +46,12 @@ class Graphene(Compound):
             angles=[90, 90, 90],
             lattice_points={"A": points},
         )
-        carbon = Compound(name="C")
+        carbon = Compound(name="C", element="C")
         layers = lattice.populate(
             compound_dict={"A": carbon}, x=x_repeat, y=y_repeat, z=n_layers
         )
-        self.add(layers)
-        self.freud_generate_bonds("C", "C", dmin=0.14, dmax=0.145)
+        surface.add(layers)
+        surface.freud_generate_bonds("C", "C", dmin=0.14, dmax=0.145)
+        super(Graphene, self).__init__(
+            compound=surface, num_mols=1, force_field=force_field
+        )
