### 2023-11-14 57bcfdf
added questions guideline to contribution guideline
