### 2023-11-14 #5989 @marjanAlbouye
@LIVazquezS  Thank you for the review and suggestions.

**Example Usage:**

- We have updated the tutorials in [this PR](https://github.com/cmelab/flowerMD/pull/95), providing additional descriptions and guidelines on creating subclasses for force fields and molecules. Tutorial 2 includes code examples for using both the Pack and Lattice classes.

- Direct references to the source code have been removed. Instead, we've added links to our [documentation](https://flowermd.readthedocs.io/en/latest/), which comprehensively covers details of classes and functions.

- A link to the `.xml` forcefield format has been added to Section 2 of Tutorial 2.

- We have initiated the development of the surface wetting module for flowerMD. You can track the progress in [this PR](https://github.com/cmelab/flowerMD/pull/96). Upon completion of implementation and testing, we'll provide a tutorial explaining the new functionalities.

**Community Guidelines:**
- A new section has been added to our [contribution guidelines ](https://github.com/cmelab/flowerMD/blob/main/CONTRIBUTING.md)for questions and comments. This section also includes instructions for bug reports and code contributions.

**Suggested Improvements:**

- As mentioned earlier, unfortunately, many of the deprecation warnings stem from packages in flowerMD dependencies. We are actively addressing the ones caused by our code.

- [flowerMD's documentation](https://flowermd.readthedocs.io/en/latest/) includes descriptions of functions and classes. This documentation has been linked in the README and paper.

- We have already initiated the process of creating the conda package in [this PR](https://github.com/conda-forge/staged-recipes/pull/24490) on conda-forge staged-recipes. Once the package is published, we will update the installation guidelines.

**Other Questions:**

Thank you for the suggestion regarding the CG force fields. While we haven't directly worked with MARTINI or SIRAH, we will make sure to explore them. However, supporting those CG force fields is currently beyond the scope of flowerMD.
