### 2023-11-15 568c014
update class interface and add methods for surface snapshot and ff
