+++ flowermd/modules/surface_wetting.py
@@ -1,34 +1,94 @@
 """Module for simulating surface wetting."""
 import hoomd
 import unyt as u
-from jankflow.base.simulation import Simulation
-from jankflow.utils import HOOMDThermostats
+from gmso.external import to_gsd_snapshot, to_hoomd_forcefield
+
+from flowermd.base import Simulation
+from flowermd.utils import HOOMDThermostats
 
 
-class Surface:
-    """For creating an interface between two slabs.
+class SurfaceDropletCreator:
 
-    Parameters
-    ----------
-    gsd_file : str
-        Path to gsd file of the slab.
-    interface_axis : tuple
-        Axis along which the interface is to be created.
-        The slab file is duplicated and translated along this axis.
-    gap : float, required
-        Distance (in simulation units) between the two slabs at the interface.
-    wall_sigma : float
-        Sigma parameter used for the wall potential when creating the slabs.
+    def __init__(self, surface, drop_snapshot, drop_ff, drop_ref_values,
+                 box_height, gap):
+        """Create the snapshot and forces for the surface droplet simulation.
 
-    """
+        Creates a new snapshot that combines the surface and droplet snapshots
+        by adding the droplet particles on top of the surface particles with
+        the correct spacing (i.e. gap) between the two.
+        Also combines the hoomd forces from the surface and droplet simulations,
+        and adds the forces for the pair interactions between the droplet and
+        surface particles.
 
-    def __init__(self, drop_gsd_file, surface_gsd_file, gap):
-        self.drop_gsd_file = drop_gsd_file
-        self.surface_gsd_file = surface_gsd_file
+        Droplet reference values are used as the reference to scale the mass,
+        energy and length values in the new snapshot.
+
+        Parameters
+        ----------
+        surface : flowermd.library.surfaces.Graphene, required
+            The surface to place the droplet on.
+        drop_snapshot : hoomd.snapshot.Snapshot or str, required
+            A snapshot of the droplet simulation, or a path
+            to a GSD file of the droplet simulation.
+        drop_ff : List of hoomd.md.force.Force, required
+            List of HOOMD force objects used in the droplet simulation.
+        drop_ref_values : dict, required
+            Dictionary of reference values for the droplet simulation.
+        box_height : unyt.unyt_quantity or float, required
+            The height of the simulation box.
+        gap : unyt.unyt_quantity or float, required
+            The gap between the droplet and the surface.'
+
+        """
+
+        self.surface = surface
+        self.drop_snapshot = drop_snapshot
+        self.drop_ff = drop_ff
+        self.drop_ref_values = drop_ref_values
+        self.box_height = box_height
         self.gap = gap
+
+        # get surface snapshot and forces
+        self._surface_snapshot, self._surface_ref_values = \
+            self._create_surface_snapshot()
+        self._surface_forces = self._create_surface_forces()
+        # scale surface ff based on droplet simulation reference values
+        self._scaled_surface_forces = self._scale_surface_ff()
+        # scale surface snapshot based on droplet simulation reference values
+        self._scaled_surface_snapshot = self._scale_surface_snapshot()
 
     def _build(self):
         """Duplicates the slab and builds the interface."""
+        pass
+
+    def _create_surface_snapshot(self):
+        """Get the surface snapshot."""
+        snap, refs = to_gsd_snapshot(
+            top=self.surface.gmso_molecule,
+            auto_scale=True,
+            base_units=None,
+        )
+        return snap, refs
+
+    def _create_surface_forces(self, surface_r_cut=2.5):
[…]
