### 2023-11-15 d6063a4
use droplet ref values to get surface snapshot and ff
