### 2023-11-15 356a0a7
get r_cut from droplet lj forcce
