### 2023-11-15 43beef0
update run_droplet params
