### 2023-11-15 c9c7404
add function for combining lj forces
