+++ flowermd/modules/surface_wetting.py
@@ -9,9 +9,9 @@ from flowermd.utils import HOOMDThermostats
 
 
 class SurfaceDropletCreator:
-
-    def __init__(self, surface, drop_snapshot, drop_ff, drop_ref_values,
-                 box_height, gap):
+    def __init__(
+        self, surface, drop_snapshot, drop_ff, drop_ref_values, box_height, gap
+    ):
         """Create the snapshot and forces for the surface droplet simulation.
 
         Creates a new snapshot that combines the surface and droplet snapshots
@@ -41,7 +41,6 @@ class SurfaceDropletCreator:
             The gap between the droplet and the surface.'
 
         """
-
         self.surface = surface
         self.drop_snapshot = drop_snapshot
         self.drop_ff = drop_ff
@@ -51,14 +50,19 @@ class SurfaceDropletCreator:
 
         self._drop_r_cut = self._get_drop_r_cut()
         # get surface snapshot and forces
-        self._surface_snapshot, self._surface_ref_values = \
-            self._create_surface_snapshot()
+        (
+            self._surface_snapshot,
+            self._surface_ref_values,
+        ) = self._create_surface_snapshot()
         self._surface_forces = self._create_surface_forces()
 
         # get droplet and surface particle types
         self._all_particle_types = list(
-            set(self._surface_snapshot.particles.types +
-                self.drop_snapshot.particles.types))
+            set(
+                self._surface_snapshot.particles.types
+                + self.drop_snapshot.particles.types
+            )
+        )
 
         self._combined_forces = self._combine_forces()
 
@@ -89,7 +93,7 @@ class SurfaceDropletCreator:
             top=self.surface.gmso_molecule,
             r_cut=surface_r_cut,
             auto_scale=False,
-            base_units=self.drop_ref_values
+            base_units=self.drop_ref_values,
         )
         for force in ff:
             force_list.extend(ff[force])
@@ -100,15 +104,15 @@ class SurfaceDropletCreator:
         force_dict = {}
         for force in hoomd_force_list:
             if isinstance(force, hoomd.md.pair.LJ):
-                force_dict['lj'] = force
+                force_dict["lj"] = force
             elif isinstance(force, hoomd.md.long_range.pppm.Coulomb):
-                force_dict['coulomb'] = force
+                force_dict["coulomb"] = force
             elif isinstance(force, hoomd.md.bond.Bond):
-                force_dict['bond'] = force
+                force_dict["bond"] = force
             elif isinstance(force, hoomd.md.angle.Angle):
-                force_dict['angle'] = force
+                force_dict["angle"] = force
             elif isinstance(force, hoomd.md.dihedral.Dihedral):
-                force_dict['dihedral'] = force
+                force_dict["dihedral"] = force
         return force_dict
 
     def _combine_forces(self):
@@ -122,7 +126,8 @@ class SurfaceDropletCreator:
 
         # combine LJ forces
         combined_force_list.append(
-            self._combine_lj(drop_forces_dict['lj'], surface_forces_dict['lj']))
+            self._combine_lj(drop_forces_dict["lj"], surface_forces_dict["lj"])
+        )
 
         # combine Coulomb forces
 
@@ -134,9 +139,9 @@ class SurfaceDropletCreator:
 
         return combined_force_list
 
-    def _combine_lj(self, drop_lj, surface_lj, combining_rule='geometric'):
+    def _combine_lj(self, drop_lj, surface_lj, combining_rule="geometric"):
         """Combine the droplet and surface LJ forces."""
-        if combining_rule not in ['geometric', 'lorentz']:
+        if combining_rule not in ["geometric", "lorentz"]:
             raise ValueError("combining_rule must be 'geometric' or 'lorentz'")
 
         lj = drop_lj.copy()
[…]
