### 2023-11-15 d5d95a7
add methods for combining bond, angle and dihedrals
