+++ flowermd/modules/welding/__init__.py
@@ -1,0 +1,2 @@
+"""Welding module for FlowerMD."""
+from .welding import Interface, SlabSimulation, WeldSimulation
