+++ flowermd/modules/surface_wetting/surface_wetting.py
@@ -1,8 +1,8 @@
 """Module for simulating surface wetting."""
 import hoomd
-import numpy as np
 import unyt as u
 from gmso.external import to_gsd_snapshot, to_hoomd_forcefield
+from utils import combine_forces
 
 from flowermd.base import Simulation
 from flowermd.utils import HOOMDThermostats
@@ -50,13 +50,12 @@ class SurfaceDropletCreator:
         self.box_height = box_height
         self.gap = gap
 
-        self._drop_r_cut = self._get_drop_r_cut()
         # get surface snapshot and forces
         (
             self._surface_snapshot,
             self._surface_ref_values,
         ) = self._create_surface_snapshot()
-        self._surface_forces = self._create_surface_forces()
+        self.surface_ff = self._create_surface_forces()
 
         # get droplet and surface particle types
         self._all_particle_types = list(
@@ -66,18 +65,11 @@ class SurfaceDropletCreator:
             )
         )
 
-        self._combined_forces = self._combine_forces()
+        self._combined_forces = combine_forces(self.drop_ff, self.surface_ff)
 
     def _build(self):
         """Duplicates the slab and builds the interface."""
         pass
-
-    def _get_drop_r_cut(self):
-        """Find the r_cut value from the droplet LJ forces."""
-        for force in self.drop_ff:
-            if isinstance(force, hoomd.md.pair.LJ):
-                lj_r_cut = force.r_cut.values()[0]
-                return lj_r_cut
 
     def _create_surface_snapshot(self):
         """Get the surface snapshot."""
@@ -100,141 +92,6 @@ class SurfaceDropletCreator:
         for force in ff:
             force_list.extend(ff[force])
         return force_list
-
-    def _retrieve_forces(self, hoomd_force_list):
-        """Retrieve individual force objects from the hoomd force list."""
-        force_dict = {}
-        for force in hoomd_force_list:
-            if isinstance(force, hoomd.md.pair.LJ):
-                force_dict["lj"] = force
-            elif isinstance(force, hoomd.md.long_range.pppm.Coulomb):
-                force_dict["coulomb"] = force
-            elif isinstance(force, hoomd.md.bond.Harmonic):
-                force_dict["bond"] = force
-            elif isinstance(force, hoomd.md.angle.Harmonic):
-                force_dict["angle"] = force
-            elif isinstance(force, hoomd.md.dihedral.OPLS):
-                force_dict["dihedral"] = force
-        return force_dict
-
-    def _combine_forces(self):
-        """Combine the surface and droplet forces."""
-        combined_force_list = []
-
-        # retrieve individual force objects from the droplet and surface
-        # force lists
-        drop_forces_dict = self._retrieve_forces(self.drop_ff)
-        surface_forces_dict = self._retrieve_forces(self._surface_forces)
-
-        # combine LJ forces
-        combined_force_list.append(
-            self._combine_lj_forces(
-                drop_forces_dict["lj"], surface_forces_dict["lj"]
-            )
-        )
-
-        # add Coulomb forces from droplet simulation if exists
-        if drop_forces_dict.get("coulomb"):
-            combined_force_list.append(drop_forces_dict["coulomb"])
-        # combine bond forces
-        combined_force_list.append(
-            self._combine_bond_forces(
-                drop_forces_dict["bond"], surface_forces_dict["bond"]
-            )
-        )
-        # combine angle forces
-        combined_force_list.append(
-            self._combine_angle_forces(
-                drop_forces_dict["angle"], surface_forces_dict["angle"]
-            )
-        )
-        # combine dihedral forces
-        combined_force_list.append(
-            self._combine_dihedral_forces(
-                drop_forces_dict["dihedral"], surface_forces_dict["dihedral"]
-            )
-        )
-        return combined_force_list
-
[…]
