+++ flowermd/modules/__init__.py
@@ -1,2 +1,7 @@
 """flowerMD modules."""
-from .welding import Interface, SlabSimulation, WeldSimulation
+from surface_wetting import (
+    DropletSimulation,
+    SurfaceDropletCreator,
+    WettingSimulation,
+)
+from welding import Interface, SlabSimulation, WeldSimulation
