+++ flowermd/modules/surface_wetting/__init__.py
@@ -1,0 +1,6 @@
+"""Surface wetting module for FlowerMD."""
+from .surface_wetting import (
+    DropletSimulation,
+    SurfaceDropletCreator,
+    WettingSimulation,
+)
