### 2023-11-16 #5989 @chrisjonesBSU
The anaconda package for `flowermd` is done as of this morning. You can see it [here](https://anaconda.org/conda-forge/flowermd) and the conda-forge feedstock repo [here](https://github.com/conda-forge/flowermd-feedstock).

@marjanAlbouye and I looked into making a pip package as well. A big chunk of the dependencies for `flowermd` are only available on anaconda and don't have pip packages in pypi. From what we understand, this might complicate the process of making a pip package for `flowermd`. Marjan found [poetry](https://pypi.org/project/poetry/) which I think can help with that. We'll keep it a goal to have a pip package at some point, but it may be a bit before we can get to that.
