### 2023-11-16 d1fe046
add badges, update install instructions
