### 2023-11-16 ebaddbc
rename env files, update docs
