+++ environment-cpu.yml
@@ -1,21 +1,0 @@
-name: flowermd
-channels:
-  - conda-forge
-dependencies:
-  - foyer
-  - fresnel
-  - freud >=2.13.1
-  - gmso >=0.11.2
-  - gsd >=3.0
-  - hoomd=4.3=*cpu*
-  - mbuild >=0.16.4
-  - numpy
-  - openbabel
-  - pip
-  - py3Dmol
-  - pytest
-  - pytest-cov
-  - python >=3.10
-  - fresnel >=0.13.5
-  - cmeutils >=1.0
-  - grits >=0.3.0
