+++ flowermd/modules/surface_wetting/utils.py
@@ -1,10 +1,12 @@
 """Utility functions for the surface wetting module."""
 
+import copy
+
 import hoomd
 import numpy as np
 
 
-def combine_forces(drop_forces, surface_forces):
+def combine_forces(drop_forces, surface_forces, drop_ptypes, surface_ptypes):
     """Combine the surface and droplet forces."""
     combined_force_list = []
 
@@ -15,7 +17,12 @@ def combine_forces(drop_forces, surface_forces):
 
     # combine LJ forces
     combined_force_list.append(
-        combine_lj_forces(drop_forces_dict["lj"], surface_forces_dict["lj"])
+        combine_lj_forces(
+            drop_forces_dict["lj"],
+            surface_forces_dict["lj"],
+            drop_ptypes,
+            surface_ptypes,
+        )
     )
 
     # add Coulomb forces from droplet simulation if exists
@@ -66,23 +73,27 @@ def combine_lj_forces(
     if combining_rule not in ["geometric", "lorentz"]:
         raise ValueError("combining_rule must be 'geometric' or 'lorentz'")
 
-    lj = drop_lj.copy()
+    lj = hoomd.md.pair.LJ(nlist=drop_lj.nlist)
+
+    # add droplet LJ parameters
+    for drop_pair, drop_params in drop_lj.params.items():
+        lj.params[drop_pair] = drop_params
+        lj.r_cut[drop_pair] = drop_lj.r_cut[drop_pair]
 
     # add the surface LJ parameters to the droplet LJ parameters
-    for k, v in surface_lj.params.items():
-        if k not in lj.params.keys():
-            lj.params[k] = v
-    for k, v in surface_lj.r_cut.items():
-        if k not in lj.r_cut.keys():
-            lj.r_cut[k] = v
+    for surface_pair, surface_params in surface_lj.params.items():
+        if surface_pair not in list(lj.params.keys()):
+            lj.params[surface_pair] = surface_params
+            lj.r_cut[surface_pair] = surface_lj.r_cut[surface_pair]
 
     # find new pairs and add them to the droplet LJ pairs
+    r_cut = list(drop_lj.r_cut.values())[0]
     for drop_ptype in drop_ptypes:
         for surface_ptype in surface_ptypes:
-            if (drop_ptype, surface_ptype) not in lj.params.keys() and (
+            if (drop_ptype, surface_ptype) not in list(lj.params.keys()) and (
                 surface_ptype,
                 drop_ptype,
-            ) not in lj.params.keys():
+            ) not in list(lj.params.keys()):
                 epsilon = np.sqrt(
                     drop_lj.params[(drop_ptype, drop_ptype)]["epsilon"]
                     * surface_lj.params[(surface_ptype, surface_ptype)][
@@ -109,13 +120,13 @@ def combine_lj_forces(
                     "sigma": sigma,
                     "epsilon": epsilon,
                 }
-                lj.r_cut[(drop_ptype, surface_ptype)] = lj.r_cut.values()[0]
+                lj.r_cut[(drop_ptype, surface_ptype)] = r_cut
     return lj
 
 
 def combine_bond_forces(drop_bond, surface_bond):
     """Combine the droplet and surface bond forces."""
-    bond = drop_bond.copy()
+    bond = copy.deepcopy(drop_bond)
     # add the surface bond parameters to the droplet bond parameters
     for k, v in surface_bond.params.items():
         if k not in bond.params.keys():
@@ -125,7 +136,7 @@ def combine_bond_forces(drop_bond, surface_bond):
 
 def combine_angle_forces(drop_angle, surface_angle):
     """Combine the droplet and surface angle forces."""
-    angle = drop_angle.copy()
+    angle = copy.deepcopy(drop_angle)
[…]
