### 2023-11-20 3d50568
update combine forces utils
