+++ flowermd/modules/surface_wetting/utils.py
@@ -1,7 +1,5 @@
 """Utility functions for the surface wetting module."""
 
-import copy
-
 import hoomd
 import numpy as np
 
@@ -12,12 +10,12 @@ def combine_forces(drop_forces, surface_forces, drop_ptypes, surface_ptypes):
 
     # retrieve individual force objects from the droplet and surface
     # force lists
-    drop_forces_dict = retrieve_forces(drop_forces)
-    surface_forces_dict = retrieve_forces(surface_forces)
+    drop_forces_dict = _retrieve_all_forces(drop_forces)
+    surface_forces_dict = _retrieve_all_forces(surface_forces)
 
     # combine LJ forces
     combined_force_list.append(
-        combine_lj_forces(
+        _combine_lj_forces(
             drop_forces_dict["lj"],
             surface_forces_dict["lj"],
             drop_ptypes,
@@ -25,31 +23,39 @@ def combine_forces(drop_forces, surface_forces, drop_ptypes, surface_ptypes):
         )
     )
 
-    # add Coulomb forces from droplet simulation if exists
+    # add Coulomb forces if exists
     if drop_forces_dict.get("coulomb"):
         combined_force_list.append(drop_forces_dict["coulomb"])
+    elif surface_forces_dict.get("coulomb"):
+        combined_force_list.append(surface_forces_dict["coulomb"])
     # combine bond forces
     combined_force_list.append(
-        combine_bond_forces(
-            drop_forces_dict["bond"], surface_forces_dict["bond"]
+        _merge_force_parameters(
+            hoomd.md.bond.Harmonic,
+            drop_forces_dict["bond"],
+            surface_forces_dict["bond"],
         )
     )
     # combine angle forces
     combined_force_list.append(
-        combine_angle_forces(
-            drop_forces_dict["angle"], surface_forces_dict["angle"]
+        _merge_force_parameters(
+            hoomd.md.angle.Harmonic,
+            drop_forces_dict["angle"],
+            surface_forces_dict["angle"],
         )
     )
     # combine dihedral forces
     combined_force_list.append(
-        combine_dihedral_forces(
-            drop_forces_dict["dihedral"], surface_forces_dict["dihedral"]
+        _merge_force_parameters(
+            hoomd.md.dihedral.OPLS,
+            drop_forces_dict["dihedral"],
+            surface_forces_dict["dihedral"],
         )
     )
     return combined_force_list
 
 
-def retrieve_forces(hoomd_force_list):
+def _retrieve_all_forces(hoomd_force_list):
     """Retrieve individual force objects from the hoomd force list."""
     force_dict = {}
     for force in hoomd_force_list:
@@ -66,7 +72,7 @@ def retrieve_forces(hoomd_force_list):
     return force_dict
 
 
-def combine_lj_forces(
+def _combine_lj_forces(
     drop_lj, surface_lj, drop_ptypes, surface_ptypes, combining_rule="geometric"
 ):
     """Combine the droplet and surface LJ forces."""
@@ -76,17 +82,17 @@ def combine_lj_forces(
     lj = hoomd.md.pair.LJ(nlist=drop_lj.nlist)
 
     # add droplet LJ parameters
-    for drop_pair, drop_params in drop_lj.params.items():
-        lj.params[drop_pair] = drop_params
-        lj.r_cut[drop_pair] = drop_lj.r_cut[drop_pair]
+    for k, v in drop_lj.params.items():
+        lj.params[k] = v
+        lj.r_cut[k] = drop_lj.r_cut[k]
 
-    # add the surface LJ parameters to the droplet LJ parameters
-    for surface_pair, surface_params in surface_lj.params.items():
-        if surface_pair not in list(lj.params.keys()):
-            lj.params[surface_pair] = surface_params
[…]
