### 2023-11-20 2d5450a
get wetting box, update positions
