+++ flowermd/library/surfaces.py
@@ -45,7 +45,12 @@ class Graphene(Molecule):
     ):
         surface = mb.Compound(periodicity=periodicity)
         spacings = [0.425, 0.246, 0.35]
-        points = [[1 / 6, 0, 0], [1 / 2, 0, 0], [0, 0.5, 0], [2 / 3, 1 / 2, 0]]
+        points = [
+                [1 / 6, 0, 0],
+                [1 / 2, 0, 0],
+                [0, 1 / 2, 0],
+                [2 / 3, 1 / 2, 0]
+        ]
         lattice = Lattice(
             lattice_spacing=spacings,
             angles=[90, 90, 90],
