+++ flowermd/modules/surface_wetting/surface_wetting.py
@@ -5,8 +5,8 @@ import numpy as np
 import unyt as u
 from gmso.external import to_gsd_snapshot, to_hoomd_forcefield
 
-from flowermd.modules.surface_wetting.utils import combine_forces
 from flowermd.base import Simulation
+from flowermd.modules.surface_wetting.utils import combine_forces
 from flowermd.utils import HOOMDThermostats
 
 
