### 2023-11-21 12dc650
load snapshot from gsd file, scale gap and box_height, fix np.concatenate calls
