### 2023-11-21 95437b7
fix grabbing LJ params from surface FF, clean up setting new params
