### 2023-11-21 1e08eed
change z_shift, add properties
