### 2023-11-21 5e827ca
set integrate group for WettingSimulation
