+++ flowermd/tests/modules/test_surface_wetting.py
@@ -1,0 +1,18 @@
+# from flowermd.modules.surface_wetting import (
+#     DropletSimulation,
+#     InterfaceBuilder,
+#     WettingSimulation,
+# )
+from flowermd.tests.base_test import BaseTest
+
+
+class TestDropletSimulation(BaseTest):
+    pass
+
+
+class TestInterfaceBuilder(BaseTest):
+    pass
+
+
+class TestWettingSimulation(BaseTest):
+    pass
