### 2023-11-21 a50ff40
move surface snap and ff to its class
