+++ flowermd/modules/surface_wetting/surface_wetting.py
@@ -3,7 +3,6 @@ import gsd.hoomd
 import hoomd
 import numpy as np
 import unyt as u
-from gmso.external import to_gsd_snapshot, to_hoomd_forcefield
 
 from flowermd.base import Simulation
 from flowermd.modules.surface_wetting.utils import combine_forces
@@ -82,24 +81,34 @@ class InterfaceBuilder:
     """Builds an interface with droplet on top of a surface."""
 
     def __init__(
-        self, surface, drop_snapshot, drop_ff, drop_ref_values, box_height, gap
+        self,
+        surface_snapshot,
+        surface_ff,
+        drop_snapshot,
+        drop_ff,
+        drop_ref_values,
+        box_height,
+        gap,
     ):
         """Create the snapshot and forces for the surface droplet simulation.
 
         Creates a new snapshot that combines the surface and droplet snapshots
-        by adding the droplet particles on top of the surface particles with
+        by putting the droplet particles on top of the surface particles with
         the correct spacing (i.e. gap) between the two.
         Also combines the hoomd forces from the surface and droplet simulations,
-        and adds the forces for the pair interactions between the droplet and
-        surface particles.
+        and adds the forces for the new pair interactions between the droplet
+        and surface particles.
 
         Droplet reference values are used as the reference to scale the mass,
         energy and length values in the new snapshot.
 
         Parameters
         ----------
-        surface : flowermd.library.surfaces.Graphene, required
-            The surface to place the droplet on.
+        surface_snapshot : hoomd.snapshot.Snapshot or str, required
+            A snapshot of the surface simulation, or a path to a GSD file
+            of the surface simulation.
+        surface_ff : List of hoomd.md.force.Force, required
+            List of HOOMD force objects used in the surface simulation.
         drop_snapshot : hoomd.snapshot.Snapshot or str, required
             A snapshot of the droplet simulation, or a path
             to a GSD file of the droplet simulation.
@@ -113,23 +122,21 @@ class InterfaceBuilder:
             The gap between the droplet and the surface.'
 
         """
-        self.surface = surface
         if isinstance(drop_snapshot, str):
             drop_snapshot = gsd.hoomd.open(drop_snapshot)[-1]
+        if isinstance(surface_snapshot, str):
+            surface_snapshot = gsd.hoomd.open(surface_snapshot)[-1]
+        self.surface_ff = surface_ff
+        self.surface_snapshot = surface_snapshot
         self.drop_snapshot = drop_snapshot
         self.drop_ff = drop_ff
         self.drop_ref_values = drop_ref_values
         self.box_height = box_height / drop_ref_values["length"]
         self.gap = gap / drop_ref_values["length"]
 
-        # get surface snapshot and forces
-        (
-            self.surface_snapshot,
-            self.surface_ref_values,
-        ) = self._create_surface_snapshot()
         self._surface_n = self.surface_snapshot.particles.N
         self._drop_n = self.drop_snapshot.particles.N
-        self.surface_ff = self._create_surface_forces()
+
         # get snapshot of the combined system
         if set(self.surface_snapshot.particles.types).intersection(
             set(self.drop_snapshot.particles.types)
@@ -324,28 +331,6 @@ class InterfaceBuilder:
         drop_pos[:, 2] -= z_shift
         wetting_pos = np.concatenate((surface_pos, drop_pos), axis=0)
         return wetting_pos
-
-    def _create_surface_snapshot(self):
-        """Get the surface snapshot."""
-        snap, refs = to_gsd_snapshot(
-            top=self.surface.gmso_molecule,
-            auto_scale=False,
-            base_units=self.drop_ref_values,
-        )
-        return snap, refs
-
[…]
