+++ flowermd/modules/__init__.py
@@ -1,7 +1,0 @@
-"""flowerMD modules."""
-from .surface_wetting import (
-    DropletSimulation,
-    SurfaceDropletCreator,
-    WettingSimulation,
-)
-from .welding import Interface, SlabSimulation, WeldSimulation
