### 2023-11-21 05041a9
add droplet sim tests
