### 2023-11-21 c3361e6
fixes to the fix_surfaces functionality
