+++ flowermd/modules/surface_wetting/surface_wetting.py
@@ -410,16 +410,16 @@ class WettingSimulation(Simulation):
         if self.fix_surface:
             snapshot = self.state.get_snapshot()
             droplet_types = np.array(
-                [i for i in snapshot if not i.startswith("_")]
+                [i for i in snapshot.particles.types if not i.startswith("_")]
             )
-            droplet_type_idx = np.where(
-                np.array(snapshot.particles.types) == droplet_types
-            )[0]
-            surface_particle_idx = np.where(
-                np.array(snapshot.particles.typeid) == droplet_type_idx
-            )[0]
+            droplet_type_idx = np.searchsorted(
+                snapshot.particles.types, droplet_types
+            )
+            droplet_particle_idx = np.searchsorted(
+                snapshot.particles.typeid, droplet_type_idx
+            )
             self.integrate_group = hoomd.filter.Tags(
-                surface_particle_idx.astype(np.uint32)
+                droplet_particle_idx.astype(np.uint32)
             )
         else:
             self.integrate_group = hoomd.filter.All()
