### 2023-11-21 be330b7
fix setting integrate group for surface
