### 2023-11-22 #5989 @lucydot
A very nicely written JOSS paper @marjanAlbouye and @chrisjonesBSU 👍
