### 2023-11-22 #5989 @lucydot
I see there is another package "MoSDeF" which is cited in your paper. It may be worth highlighting how flowerMD is related to / distinct from MoSDeF in the statement of need - it appears there is crossover?
