### 2023-11-22 fdf60cd
change graphene to inherit from system
