### 2023-11-22 5f4d2e4
move apply_FF params outisde of init, add test to apply a FF
