### 2023-11-27 c76cbd6
fix some unit test errors
