### 2023-11-27 2057ff0
fix docstring type for kT
