### 2023-11-27 0b3cb3e
add shrink density parameter to run_droplet;add docstring
