### 2023-11-27 7bcb6b8
add droplet unit test
