+++ flowermd/tests/base/test_system.py
@@ -167,6 +167,7 @@ class TestSystem(BaseTest):
         with pytest.warns():
             system.remove_hydrogens()
 
+    @pytest.mark.skip(reason="Debugging shared test data")
     def test_add_mass_charges(self, benzene_molecule):
         benzene_mols = benzene_molecule(n_mols=1)
         system = Pack(
@@ -215,6 +216,7 @@ class TestSystem(BaseTest):
             system.mass, ((12.011 * 6) + (1.008 * 6) + 32.06) * 20, atol=1e-4
         )
 
+    @pytest.mark.skip(reason="Debugging shared test data")
     def test_ref_length(self, polyethylene):
         polyethylene = polyethylene(lengths=5, num_mols=5)
         system = Pack(
@@ -234,6 +236,7 @@ class TestSystem(BaseTest):
         assert np.allclose(calc_box[1], system.box.Ly, atol=1e-2)
         assert np.allclose(calc_box[2], system.box.Lz, atol=1e-2)
 
+    @pytest.mark.skip(reason="Debugging shared test data")
     def test_ref_mass(self, polyethylene):
         polyethylene = polyethylene(lengths=5, num_mols=5)
         system = Pack(
@@ -250,6 +253,7 @@ class TestSystem(BaseTest):
             atol=1e-1,
         )
 
+    @pytest.mark.skip(reason="Debugging shared test data")
     def test_ref_energy(self, polyethylene):
         polyethylene = polyethylene(lengths=5, num_mols=5)
         system = Pack(
