+++ flowermd/tests/base_test.py
@@ -267,3 +267,11 @@ class BaseTest:
             },
         )
         return ff.hoomd_forces
+
+    @pytest.fixture()
+    def polyethylene_droplet(self):
+        return os.path.join(ASSETS_DIR, "polyethylene_droplet.gsd")
+
+    @pytest.fixture()
+    def graphene_snapshot(self):
+        return os.path.join(ASSETS_DIR, "graphene_snapshot.gsd")
