### 2023-11-27 1c30afe
add wetting sim test
