+++ flowermd/tests/base/test_system.py
@@ -167,13 +167,9 @@ class TestSystem(BaseTest):
         with pytest.warns():
             system.remove_hydrogens()
 
-    @pytest.mark.skip(reason="Debugging shared test data")
     def test_add_mass_charges(self, benzene_molecule):
         benzene_mols = benzene_molecule(n_mols=1)
-        system = Pack(
-            molecules=[benzene_mols],
-            density=0.8,
-        )
+        system = Pack(molecules=[benzene_mols], density=0.8, base_units=dict())
         system.apply_forcefield(
             r_cut=2.5,
             force_field=OPLS_AA(),
@@ -216,13 +212,9 @@ class TestSystem(BaseTest):
             system.mass, ((12.011 * 6) + (1.008 * 6) + 32.06) * 20, atol=1e-4
         )
 
-    @pytest.mark.skip(reason="Debugging shared test data")
     def test_ref_length(self, polyethylene):
         polyethylene = polyethylene(lengths=5, num_mols=5)
-        system = Pack(
-            molecules=[polyethylene],
-            density=1.0,
-        )
+        system = Pack(molecules=[polyethylene], density=1.0, base_units=dict())
         system.apply_forcefield(
             r_cut=2.5, force_field=[OPLS_AA()], auto_scale=True
         )
@@ -236,13 +228,9 @@ class TestSystem(BaseTest):
         assert np.allclose(calc_box[1], system.box.Ly, atol=1e-2)
         assert np.allclose(calc_box[2], system.box.Lz, atol=1e-2)
 
-    @pytest.mark.skip(reason="Debugging shared test data")
     def test_ref_mass(self, polyethylene):
         polyethylene = polyethylene(lengths=5, num_mols=5)
-        system = Pack(
-            molecules=[polyethylene],
-            density=1.0,
-        )
+        system = Pack(molecules=[polyethylene], density=1.0, base_units=dict())
         system.apply_forcefield(
             r_cut=2.5, force_field=[OPLS_AA()], auto_scale=True
         )
@@ -253,13 +241,9 @@ class TestSystem(BaseTest):
             atol=1e-1,
         )
 
-    @pytest.mark.skip(reason="Debugging shared test data")
     def test_ref_energy(self, polyethylene):
         polyethylene = polyethylene(lengths=5, num_mols=5)
-        system = Pack(
-            molecules=[polyethylene],
-            density=1.0,
-        )
+        system = Pack(molecules=[polyethylene], density=1.0, base_units=dict())
         system.apply_forcefield(
             r_cut=2.5, force_field=[OPLS_AA()], auto_scale=True
         )
