### 2023-11-28 #5989 @marjanAlbouye
@lucydot

We completed all of the above author tasks and released the software.

Link to the archive release on Zenodo:

[https://zenodo.org/records/10211282](https://zenodo.org/records/10211282)

DOI: 10.5281/zenodo.10211282

We also merged the surface wetting module ([this PR](https://github.com/cmelab/flowerMD/pull/96.)) to our repository and released `flowermd` version `1.1.0`. A new tutorial is added for this module that covers the details of it. We also added a few sentences to the paper mentioning this new module.
