### 2023-11-29 #5989 @chrisjonesBSU
@lucydot one change to Marjan's last comment after fixing a quick issue in the repo's `setup.py` file

https://zenodo.org/records/10215501
DOI: 10.5281/zenodo.10215501
version: 1.1.1

Thanks!
