### 2023-11-30 f490cfd
add class for ellipsoid chain FF
