### 2023-12-01 97aff78
add FF class to library
