+++ flowermd/library/__init__.py
@@ -15,8 +15,8 @@ from .polymers import (
     PEEK,
     PEKK,
     PPS,
-    LJChain,
     EllipsoidChain,
+    LJChain,
     PEKK_meta,
     PEKK_para,
     PolyEthylene,
