### 2023-12-04 #5989 @marjanAlbouye
> > We'll take a closer look at the statement of need and see if we can be more clear about how MoSDeF is involved, and how flowerMD extends on top of MoSDeF.
>
> Before I recommend accept, did you want to adjust the statement of need @chrisjonesBSU @marjanAlbouye ?

@lucydot  We made a small change to the last paragraph of the statement of the need, mentioning the MosDeF and HOOMD packages. Do we need to archive the package on Zenodo again?

Thanks!
