### 2023-12-07 4af2795
update particle mass, add bead particle types sequence
