### 2023-12-07 fb0708f
add func to create rigid frame
