+++ flowermd/utils/rigid_body.py
@@ -1,8 +1,8 @@
 import gsd
 import gsd.hoomd
+import hoomd
 import numpy as np
 from cmeutils.geometry import moit
-import hoomd
 
 
 def _get_com_mass_pos_moi(snapshot, rigid_const_idx):
@@ -12,16 +12,22 @@ def _get_com_mass_pos_moi(snapshot, rigid_const_idx):
     for idx in rigid_const_idx:
         total_mass = np.sum(snapshot.particles.mass[idx])
         com_mass.append(total_mass)
-        com = np.sum(
-            snapshot.particles.position[idx]
-            * snapshot.particles.mass[idx, np.newaxis],
-            axis=0,
-        ) / total_mass
+        com = (
+            np.sum(
+                snapshot.particles.position[idx]
+                * snapshot.particles.mass[idx, np.newaxis],
+                axis=0,
+            )
+            / total_mass
+        )
         com_position.append(com)
         com_moi.append(
-            moit(points=snapshot.particles.position[idx],
-                 masses=snapshot.particles.mass[idx],
-                 center=com))
+            moit(
+                points=snapshot.particles.position[idx],
+                masses=snapshot.particles.mass[idx],
+                center=com,
+            )
+        )
     return com_mass, com_position, com_moi
 
 
@@ -44,26 +50,30 @@ def create_rigid_body(snapshot, bead_constituents_types):
     """
     # find typeid sequence of the constituent particles types in a rigid bead
     p_types = np.array(snapshot.particles.types)
-    constituent_type_ids = \
-        np.where(p_types[:, None] == bead_constituents_types)[0]
+    constituent_type_ids = np.where(
+        p_types[:, None] == bead_constituents_types
+    )[0]
 
     # find indices that matches the constituent particle types
     typeids = snapshot.particles.typeid
     bead_len = len(bead_constituents_types)
     matches = np.where((typeids.reshape(-1, bead_len) == constituent_type_ids))
     if len(matches[0]) == 0:
-        raise ValueError("No matches found between particle types in the system"
-                         " and bead constituents particles")
+        raise ValueError(
+            "No matches found between particle types in the system"
+            " and bead constituents particles"
+        )
     rigid_const_idx = (matches[0] * bead_len + matches[1]).reshape(-1, bead_len)
 
     n_rigid = rigid_const_idx.shape[0]  # number of rigid bodies
 
     # calculate center of mass and its position for each rigid body
-    com_mass, com_position, com_moi = _get_com_mass_pos_moi(snapshot,
-                                                            rigid_const_idx)
+    com_mass, com_position, com_moi = _get_com_mass_pos_moi(
+        snapshot, rigid_const_idx
+    )
 
     rigid_frame = gsd.hoomd.Frame()
-    rigid_frame.particles.types = ['R'] + snapshot.particles.types
+    rigid_frame.particles.types = ["R"] + snapshot.particles.types
     rigid_frame.particles.N = n_rigid
     rigid_frame.particles.typeid = [0] * n_rigid
     rigid_frame.particles.mass = com_mass
@@ -73,11 +83,12 @@ def create_rigid_body(snapshot, bead_constituents_types):
 
     # find local coordinates of the particles in the first rigid body
     # only need to find the local coordinates for the first rigid body
-    local_coords = snapshot.particles.position[rigid_const_idx[0]] - \
-                   com_position[0]
+    local_coords = (
+        snapshot.particles.position[rigid_const_idx[0]] - com_position[0]
+    )
 
     rigid_constrain = hoomd.md.constrain.Rigid()
-    rigid_constrain.body['rigid'] = {
+    rigid_constrain.body["rigid"] = {
[…]
