### 2023-12-07 e8247bc
fix numpy bug, add box to config
