### 2023-12-08 cb70367
add fix_orientation variable to pack
