### 2023-12-08 fe25155
remove create_rigid_body from sim
