### 2023-12-08 0c09d82
add body tags to rigid frame
