### 2023-12-08 963b8cf
add new functions to __init__
