+++ flowermd/utils/__init__.py
@@ -1,6 +1,7 @@
 from .actions import *
 from .base_types import HOOMDThermostats
 from .ff_utils import xml_to_gmso_ff
+from .rigid_body import create_rigid_body
 from .utils import (
     calculate_box_length,
     check_return_iterable,
