+++ flowermd/tests/library/test_forcefields.py
@@ -11,6 +11,7 @@ from flowermd.library import (
     OPLS_AA_DIMETHYLETHER,
     OPLS_AA_PPS,
     BeadSpring,
+    EllipsoidForcefield,
     FF_from_file,
     TableForcefield,
 )
@@ -92,6 +93,29 @@ class TestForceFields:
             assert ff.hoomd_forces[3].params[param]["k"] == 100
             assert ff.hoomd_forces[3].params[param]["d"] == -1
             assert ff.hoomd_forces[3].params[param]["n"] == 1
+
+    def test_ellipsoid_ff(self):
+        ellipsoid_ff = EllipsoidForcefield(
+            epsilon=1.0,
+            lperp=0.5,
+            lpar=1.0,
+            bead_length=1,
+            r_cut=3,
+            bond_k=500,
+            bond_r0=0.1,
+            angle_k=100,
+            angle_theta0=1.57,
+        )
+        assert len(ellipsoid_ff.hoomd_forces) == 3
+        assert isinstance(
+            ellipsoid_ff.hoomd_forces[-1], hoomd.md.pair.aniso.GayBerne
+        )
+        assert ("R", "R") in list(
+            dict(ellipsoid_ff.hoomd_forces[-1].params).keys()
+        )
+        assert ("B", "R") in list(
+            dict(ellipsoid_ff.hoomd_forces[-1].params).keys()
+        )
 
 
 class TestTableForcefield:
