### 2023-12-08 ca67661
move private func to bottom
