### 2023-12-08 2b49384
add example code to docstring
