### 2023-12-11 418bc29
replace bead_length with lpar in EllipsoidChain class
