### 2023-12-11 e635f13
add rigid body handling of mass
