+++ flowermd/base/system.py
@@ -701,6 +701,8 @@ class Pack(System):
         The factor by which to expand the box for packing.
     edge : float, default 0.2
         The space (nm) between the edge of the box and the molecules.
+    overlap : float, default 0.2
+        Minimum separation (nm) between particles of different molecules.
 
 
     .. warning::
@@ -729,10 +731,12 @@ class Pack(System):
         base_units=dict(),
         packing_expand_factor=5,
         edge=0.2,
+        overlap=0.2,
         fix_orientation=False,
     ):
         self.packing_expand_factor = packing_expand_factor
         self.edge = edge
+        self.overlap = overlap
         self.fix_orientation = fix_orientation
         super(Pack, self).__init__(
             molecules=molecules,
@@ -746,7 +750,7 @@ class Pack(System):
             compound=self.all_molecules,
             n_compounds=[1 for i in self.all_molecules],
             box=list(self._target_box * self.packing_expand_factor),
-            overlap=0.2,
+            overlap=self.overlap,
             edge=self.edge,
             fix_orientation=self.fix_orientation,
         )
