### 2023-12-12 1e9cd6f
remove bead_length from Ellipsoid FF class, fix unit test
