### 2023-12-12 74eae9d
fix rigid unit test
