### 2023-12-12 9d963fe
add 2 unit tests for rigid simulations
