### 2023-12-13 375f40c
fix check for body tags in mass, add assertion for mass
