### 2023-12-13 3b382c6
use >= when checking for ascending next body tag
