### 2023-12-15 #5989 @Kevin-Mattheus-Moerman
@marjanAlbouye as AEiC I will now help to process final steps. I have check the paper, this review, the archive link, and your repository. Most seems in order, however, I do have the below point that requires your attention.

On the paper:
- [ ] It looks like this should be added as the DOI for the below reference: [10.1371/journal.pcbi.1005659](https://doi.org/10.1371/journal.pcbi.1005659)
```
Eastman, P., Swails, J., Chodera, J. D., McGibbon, R. T., Zhao, Y., Beauchamp, K. A., Wang,
L.-P., Simmonett, A. C., Harrigan, M. P., Stern, C. D., & others. (2017). OpenMM
7: Rapid development of high performance algorithms for molecular dynamics. PLoS
Computational Biology, 13(7), e1005659.
```
