+++ paper/paper.bib
@@ -110,6 +110,7 @@
   title={OpenMM 7: Rapid development of high performance algorithms for molecular dynamics},
   author={Eastman, Peter and Swails, Jason and Chodera, John D and McGibbon, Robert T and Zhao, Yutong and Beauchamp, Kyle A and Wang, Lee-Ping and Simmonett, Andrew C and Harrigan, Matthew P and Stern, Chaya D and others},
   journal={PLoS computational biology},
+  doi={10.1371/journal.pcbi.1005659}
   volume={13},
   number={7},
   pages={e1005659},
